Moved class to helper
